Moved RegistrationEvent class to events folder
