+++ controllers/AccountController.php
@@ -23,36 +23,8 @@ use comyii\user\models\LoginForm;
 use comyii\user\models\RecoveryForm;
 use comyii\user\models\User;
 use comyii\user\models\SocialProfile;
+use comyii\user\events\RegistrationEvent;
 
-class RegistrationEvent extends \yii\base\Event
-{
-    /**
-     *
-     * @var Model the user model
-     */
-    public $model;
-    /**
-     * @var string|null the main view file to be rendered. If null then the default view file is used. This is used so that the view file
-     * can be changed by event handlers.
-     */
-    public $viewFile;
-    /**
-     * @var boolean the current status for the controller. This is used so that event handlers can tell the controller whether to not to continue.
-     */
-    public $error = false;
-    /**
-     * @var string|null the flash message for the controller. This is used so that event handlers can update the success messages for things like user registration.
-     */
-    public $message;
-    /**
-     * @var string the type of registration. This is used if there are multiple registration types (ie. different user types)
-     */
-    public $type;
-    public $activate;
-    public $flashType;
-    public $redirect;
-    public $isActivated = false;
-}
 
 /**
  * Account controller for authentication of various user actions.
